+++ src/defect-place/defect-place.service.ts
@@ -75,6 +75,9 @@ export class DefectPlaceService {
   }
 
   async removes(idxs: []) {
+    if (idxs.length <= 0) {
+      throw new NotFoundException('삭제할 정보가 없습니다.');
+    }
     await this.dfpRepository.createQueryBuilder()
       .delete()
       .where(" dfp_idx IN (:idxs)", { idxs: idxs })
